isAnyBlank和isAnyEmpty;
